cleaned up utils to separate components from utility functions added types for operator object to improve static type analyzis and help prevent minor bugs by tooling
