+++ frontend/src/pages/operatorBundlePage/bundlePageUtils.js
@@ -1,13 +1,7 @@
 import * as React from 'react';
-import * as _ from 'lodash-es';
-import classNames from 'classnames';
+import _ from 'lodash-es';
 import { safeDump, safeLoad } from 'js-yaml';
 
-import {
-  operatorFieldDescriptions,
-  operatorFieldPlaceholders,
-  operatorFieldValidators
-} from '../../utils/operatorDescriptors';
 import { getFieldValueError, getDefaultDescription } from '../../utils/operatorUtils';
 import { OPERATOR_DESCRIPTION_ABOUT_HEADER, OPERATOR_DESCRIPTION_PREREQUISITES_HEADER } from '../../utils/constants';
 
@@ -47,65 +41,17 @@ const sectionsFields = {
   'install-modes': 'spec.installModes'
 };
 
+/**
+ *
+ * @param {string|string[]} field
+ * @param {*} formErrors
+ */
 const renderFormError = (field, formErrors) => {
   const error = _.get(formErrors, field);
   if (!error) {
     return null;
   }
   return <div className="oh-operator-editor-form__error-block">{error}</div>;
-};
-
-const renderOperatorInput = (title, field, inputComponent, formErrors) => {
-  const formFieldClasses = classNames({
-    'oh-operator-editor-form__field': true,
-    row: true,
-    'oh-operator-editor-form__field--error': _.get(formErrors, field)
-  });
-
-  return (
-    <div className={formFieldClasses}>
-      <div className="form-group col-sm-6">
-        <label htmlFor={field}>{title}</label>
-        {inputComponent}
-        {renderFormError(field, formErrors)}
-      </div>
-      <div className="oh-operator-editor-form__description col-sm-6">{_.get(operatorFieldDescriptions, field, '')}</div>
-    </div>
-  );
-};
-
-const renderOperatorFormField = (operator, formErrors, updateOperator, commitField, title, field, fieldType) => {
-  let inputComponent;
-
-  if (fieldType === 'text-area') {
-    inputComponent = (
-      <textarea
-        id={field}
-        className="form-control"
-        rows={3}
-        {..._.get(_.get(operatorFieldValidators, field), 'props')}
-        onChange={e => updateOperator(e.target.value, field)}
-        onBlur={() => commitField(field)}
-        value={_.get(operator, field, '')}
-        placeholder={_.get(operatorFieldPlaceholders, field)}
-      />
-    );
-  } else {
-    inputComponent = (
-      <input
-        id={field}
-        className="form-control"
-        type={fieldType}
-        {..._.get(_.get(operatorFieldValidators, field), 'props')}
-        onChange={e => updateOperator(e.target.value, field)}
-        onBlur={() => commitField(field)}
-        value={_.get(operator, field, '')}
-        placeholder={_.get(operatorFieldPlaceholders, field)}
-      />
-    );
-  }
-
-  return renderOperatorInput(title, field, inputComponent, formErrors);
 };
 
 /**
@@ -285,8 +231,6 @@ const getMissingCrdUploads = (uploads, operator) => {
 
 export {
   sectionsFields,
-  renderOperatorInput,
-  renderOperatorFormField,
   renderFormError,
   EDITOR_STATUS,
   getUpdatedFormErrors,
